+++ docs/whats-new.md
@@ -1,5 +1,8 @@
 # What's new
 
+## UNRELEASED 
+
+
 ## 0.40.0 (2025-10-08)
 
 ### Feat